feat(demo): package eigdemo/tripleangledemo/twistdemo as roboticstoolbox.demo
The three demo console-scripts (pyproject.toml [project.scripts]) pointed
at roboticstoolbox.examples.*, a module path that hasn't existed since the
demos were relocated to a top-level examples/ dir outside the packaged
src/roboticstoolbox tree (wheel.packages = ["src/roboticstoolbox"]) --
every installed entry point 404'd at import time.

- Move the three scripts into src/roboticstoolbox/demo/ (new package,
  mirroring the existing bin/ pattern used by rtbtool).
- twistdemo.py: wrap its previously flat top-level code in main(), matching
  the other two.
- tripleangledemo.py/twistdemo.py: replace `while True: env.step(0)` with
  `env.run(duration=...)` -- besides enabling a --test smoke-test mode
  (bounded duration instead of running forever), this is a genuine fix on
  its own: the raw loop never noticed the browser tab closing, so closing
  it just left the process spinning forever; run() already has
  disconnect-awareness built in.
- eigdemo.py: fix `-h`/`--help`, which compared the whole sys.argv list
  against a string and so could never match; add the repo's standard
  hashbang line.
- pyproject.toml: repoint the three [project.scripts] entries; fix the
  matching coverage omit path (also stale, pointing at the same
  never-existed src/roboticstoolbox/examples/*).
